Add files via upload
Simon Says is the follow me memory game. The original electronic version had 4 coloured lights

This version uses the microBit and its 2 buttons to mimic the game but with 3 letters A, B and C (achieved by pressing the A+B buttons together)
If you get the sequence right, the sequence is lengthened.
If you get it wrong, you start again.